pyproject.toml
[tool.poetry]
name = "private-legal-assistant"
version = "0.0.2"
description = ""
authors = ["Lucas Coelho Pini de Sousa <lucas.pini@hotmail.com>"]
readme = "README.md"
packages = [{ include = "legal_assistant" }]

[tool.poetry.dependencies]
python = "^3.12"
llama-stack = "^0.2.2"
langchain-community = "^0.3.23"
pypdf = "^5.4.0"
langchain-chroma = "^0.2.3"
langchain-ollama = "^0.3.2"
torch = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"